Extract test_scatter_preserves_subclass from test_distributed.py to test_distributed_ops.py
